extracted functions from index.js, added jwt
