+++ package.json
@@ -18,6 +18,7 @@
     "bcryptjs": "^2.4.3",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
+    "jsonwebtoken": "^8.5.1",
     "knex": "^0.95.11",
     "pg": "^8.7.1",
     "supertest": "^6.1.6"
